Check that the expected number of ECs until corruption is 1/ec_corruption_p
This allows to check if the EC corruption CTMC has the same expected
number of steps until a first EC corruption occurs as a geometric
distribution of parameter ec_corruption_p.

Note that it is not a proof in general because model checking only
verifies the query for specific values assigned to the ec_corruption_p.